Rename folder from "discard" to "tree_manipulation"
