+++ crates/but-workspace/tests/workspace/main.rs
@@ -1,5 +1,5 @@
 mod commit_engine;
-mod discard;
 mod head_info;
 mod ref_metadata;
+mod tree_manipulation;
 mod utils;
